Generalized the PLS100 into a PLA device of variable amounts of inputs/outputs/terms. [Curt Coder]

+++ src/mess/drivers/c64.c
@@ -70,8 +70,8 @@ void c64_state::bankswitch(offs_t offset, offs_t va, int rw, int aec, int ba, in
 	int game = m_exp->game_r(offset, ba, rw, m_hiram);
 	int exrom = m_exp->exrom_r(offset, ba, rw, m_hiram);
 
-	UINT16 input = VA12 << 15 | VA13 << 14 | game << 13 | exrom << 12 | rw << 11 | aec << 10 | ba << 9 | A12 << 8 | A13 << 7 | A14 << 6 | A15 << 5 | m_va14 << 4 | m_charen << 3 | m_hiram << 2 | m_loram << 1 | cas;
-	UINT8 data = m_pla->read(input);
+	UINT32 input = VA12 << 15 | VA13 << 14 | game << 13 | exrom << 12 | rw << 11 | aec << 10 | ba << 9 | A12 << 8 | A13 << 7 | A14 << 6 | A15 << 5 | m_va14 << 4 | m_charen << 3 | m_hiram << 2 | m_loram << 1 | cas;
+	UINT32 data = m_pla->read(input);
 
 	*casram = BIT(data, 0);
 	*basic = BIT(data, 1);
